# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nervum API
  description: |
    API for Nervum – environment maps SaaS (visualize and manage services, databases,
    queues, and relationships). This backend powers the Nervum web app and can be
    used for integrations.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Users]
      summary: List users
      security: [{ cookieAuth: [] }]
      parameters:
        - name: organization_id
          in: query
          description: Filter by organization (optional)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Users]
      summary: Update user
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Delete user
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams:
    post:
      tags: [Teams]
      summary: Create team
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Teams]
      summary: List teams
      security: [{ cookieAuth: [] }]
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams/{id}:
    get:
      tags: [Teams]
      summary: Get team by ID
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Teams]
      summary: Update team
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Teams]
      summary: Delete team
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/user-teams:
    post:
      tags: [User Teams]
      summary: Add user to team
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserTeamCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTeam'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [User Teams]
      summary: List user-team links
      security: [{ cookieAuth: [] }]
      parameters:
        - name: user_id
          in: query
          description: Filter by user (provide user_id OR team_id)
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Filter by team (provide user_id OR team_id)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTeam'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [User Teams]
      summary: Remove user from team
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, team_id]
              properties:
                user_id:
                  type: string
                  format: uuid
                team_id:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/environments:
    post:
      tags: [Environments]
      summary: Create environment
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Environments]
      summary: List environments
      security: [{ cookieAuth: [] }]
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/environments/{id}:
    get:
      tags: [Environments]
      summary: Get environment by ID
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Environments]
      summary: Update environment
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Environments]
      summary: Delete environment
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/user-environment-access:
    post:
      tags: [User Environment Access]
      summary: Create user-environment access
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEnvironmentAccess'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvironmentAccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [User Environment Access]
      summary: List user-environment access
      security: [{ cookieAuth: [] }]
      parameters:
        - name: user_id
          in: query
          description: Filter by user (user_id OR environment_id required)
          schema:
            type: string
            format: uuid
        - name: environment_id
          in: query
          description: Filter by environment (user_id OR environment_id required)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserEnvironmentAccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/user-environment-access/{id}:
    get:
      tags: [User Environment Access]
      summary: Get user-environment access by ID
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvironmentAccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [User Environment Access]
      summary: Update user-environment access
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEnvironmentAccess'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvironmentAccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [User Environment Access]
      summary: Delete user-environment access
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nervum_session
      description: Session ID (UUID) set by login/register/accept-invitation
  parameters:
    UUIDPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request (e.g. validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, member]
        organization_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, member]
        organization_id:
          type: string
          format: uuid
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        organization_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [admin, manager, member]
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TeamResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Team'
        - type: object
          properties:
            environment_ids:
              type: array
              items:
                type: string
                format: uuid
    TeamCreate:
      type: object
      required: [organization_id, name]
      properties:
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        environment_ids:
          type: array
          items:
            type: string
            format: uuid
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        environment_ids:
          type: array
          items:
            type: string
            format: uuid
    UserTeam:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        role:
          type: string
        created_at:
          type: string
          format: date-time
    UserTeamCreate:
      type: object
      required: [user_id, team_id]
      properties:
        user_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        role:
          type: string
...
    Environment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [healthy, warning, critical]
        created_at:
          type: string
          format: date-time
    EnvironmentResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Environment'
        - type: object
          properties:
            services_count:
              type: integer
              description: Number of entities (services) in this environment
    EnvironmentCreate:
      type: object
      required: [organization_id, name]
      properties:
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [healthy, warning, critical]
    EnvironmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [healthy, warning, critical]
...
    UserEnvironmentAccess:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        role:
          type: string
        created_at:
          type: string
          format: date-time
security:
  - cookieAuth: []